Spot lights are fixed (eventually). Two sided materials are now properly implemented (they are rendered two times, once with inverted normals). Added a way to move lights, in CastorViewer, by first selecting a geometry which is on attached to the same scene node. Fixed ambient lighting.

+++ source/Core/Castor3D/Src/LightCache.cpp
@@ -133,7 +133,7 @@ namespace Castor3D
 		{
 			for ( auto l_it : m_typeSortedLights )
 			{
-				l_lightsCount->GetValue( 0 )[l_it.first] = uint32_t( l_it.second.size() );
+				l_lightsCount->GetValue( 0 )[size_t( l_it.first )] = uint32_t( l_it.second.size() );
 			}
 		}
 	}
